Optimized runtime memory consumption.
This should help to make it reliably working even with less than 400 bytes on the stack

+++ Spawner.h
@@ -14,12 +14,12 @@ direction: 0=towards start, 1=away from start
 class Spawner
 {
   public:
-    void Spawn(int pos, int rate, int speed, bool dir, long activate);  
+    void Spawn(int pos, int rate, uint8_t speed, bool dir, long activate);  
     void Kill();
     bool Alive();
     int _pos;
     int _rate;
-    int _speed;
+    uint8_t _speed;
     bool _dir;
     long _lastSpawned;
     long _activate;
@@ -27,7 +27,7 @@ class Spawner
     bool _alive;
 };
 
-void Spawner::Spawn(int pos, int rate, int speed, bool dir, long activate){
+void Spawner::Spawn(int pos, int rate, uint8_t speed, bool dir, long activate){
     _pos = pos;
     _rate = rate;
     _speed = speed;
